refactoring: split FileUpload => Mixins(eventMapping, validatorFile
