+++ packages/vue-dot/src/patterns/FileUpload/FileUpload.vue
@@ -5,7 +5,7 @@
 	-->
 	<label
 		v-ripple="!noRipple"
-		class="file-upload d-block pa-4"
+		class="vd-file-upload d-block pa-4"
 		:class="[
 			{
 				'dragover': dragover,
@@ -24,7 +24,7 @@
 			:disabled="disabled"
 			:multiple="multiple"
 			:accept="computedAccept"
-			class="file-upload-input"
+			class="vd-file-upload-input"
 			@change="inputValueChanged"
 		>
 
@@ -33,11 +33,10 @@
 			Every part is in a slot so it can be translated by the developer
 		-->
 		<slot name="placeholder">
-			<span class="file-upload-placeholder">
+			<span class="vd-file-upload-placeholder">
 				<slot name="icon">
 					<VIcon
-						size="40"
-						color="primary"
+						v-bind="options.uploadIcon"
 					>
 						{{ uploadIcon }}
 					</VIcon>
@@ -69,7 +68,7 @@
 					a specific Vuetify class in hover state
 				-->
 				<span
-					class="file-upload-btn primary white--text text-uppercase py-2 px-4 elevation-2"
+					class="vd-file-upload-btn primary white--text text-uppercase py-2 px-4 elevation-2"
 					:class="{ 'primary lighten-1': hover }"
 					@mouseover="hover = true"
 					@mouseleave="hover = false"
@@ -101,36 +100,16 @@
 	import Component, { mixins } from 'vue-class-component';
 
 	import { locales } from './locales';
+	import { config } from './config';
 
-	import { getFileExtension } from '../../functions/getFileExtension';
-	import { calcHumanFileSize } from '../../functions/calcHumanFileSize';
+	import { EventsMapping } from './mixins/eventsMapping';
+
+	import { customizable } from '../../mixins/customizable';
 
 	import { mdiCloudUpload } from '@mdi/js';
 
-	import { Refs } from '../../types';
-	import { ErrorEvent } from './types';
-	import { ErrorCodes } from './errorCodes';
-
-	interface HTMLInputEvent extends Event {
-		target: HTMLInputElement & EventTarget;
-	}
-
 	const Props = Vue.extend({
 		props: {
-			/** Allow multiple files */
-			multiple: {
-				type: Boolean,
-				default: false
-			},
-			/**
-			 * The v-model value
-			 * (Allow File as type because on single
-			 * mode the v-model isn't an array)
-			 */
-			value: {
-				type: [Array, Object, File],
-				default: () => []
-			},
 			/** Disable v-ripple on the component */
 			noRipple: {
 				type: Boolean,
@@ -140,42 +119,11 @@
 			disabled: {
 				type: Boolean,
 				default: false
-			},
-			/** Maximum size in bytes per file */
-			fileSizeMax: {
-				type: Number,
-				default: 4096 * 1024 // Default 4MB
-			},
-			/** The size units used in the template for i18n */
-			fileSizeUnits: {
-				type: Array as PropType<string[]>,
-				default: () => locales.fileSizeUnits
-			},
-			/** The allowed file extensions */
-			allowedExtensions: {
-				type: Array as PropType<string[]>,
-				default: () => [
-					'pdf',
-					'jpg',
-					'jpeg',
-					'png'
-				]
-			},
-			/**
-			 * The accept attribute of <input type="file">
-			 * See https://developer.mozilla.org/fr/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Complete_list_of_MIME_types
-			 *
-			 * This property is not required, by default it will be computed
-			 * based on allowedExtensions
-			 */
-			accept: {
-				type: String,
-				default: undefined
 			}
 		}
 	});
 
-	const MixinsDeclaration = mixins(Props);
+	const MixinsDeclaration = mixins(Props, customizable(config), EventsMapping);
 
 	/**
 	 * FileUpload is a component that enhance the default HTML
@@ -188,10 +136,6 @@
 		}
 	})
 	export default class FileUpload extends MixinsDeclaration {
-		$refs!: Refs<{
-			vdInputEl: HTMLInputElement;
-		}>;
-
 		// Locales
 		locales = locales;
 
@@ -199,14 +143,13 @@
 		uploadIcon = mdiCloudUpload;
 
 		/** The list of accepted files */
-		files: File[] = [];
+		files!: File[];
 		/** For specific styles on hover */
-		hover = false;
+		hover!: boolean;
 		/** For specific styles on hover with a file (dragover) */
-		dragover = false;
+		dragover!: boolean;
 		/** Used to not trigger "success" events when there is an error */
-		error = false;
-
+		error!:boolean;
 		/**
 		 * Get the different colors
 		 * depending on theme (light or dark)
@@ -220,181 +163,11 @@
 				info: 'grey--text ' + (dark ? 'text--lighten-1' : 'text--darken-1')
 			};
 		}
-
-		/** Reset self state to initial */
-		selfReset() {
-			this.dragover = false;
-			this.error = false;
-
-			// Clear previous files
-			this.files = [];
-		}
-
-		/** Validate the file (size & extension) */
-		validateFile(file: File) {
-			// Maximum size
-			if (file.size >= this.fileSizeMax) {
-				this.error = true;
-
-				this.$emit('error', {
-					file,
-					code: ErrorCodes.FILE_TOO_LARGE
-				});
-
-				return false;
-			}
-
-			const fileExt = getFileExtension(file.name);
-
-			// Extension
-			if (!this.allowedExtensions.includes(fileExt)) {
-				this.error = true;
-
-				this.$emit('error', {
-					file,
-					code: ErrorCodes.FILE_EXTENSION_NOT_ALLOWED
-				});
-
-				return false;
-			}
-
-			// Only add the file if valid
-			this.files.push(file);
-
-			return true;
-		}
-
-		/** Computed extensions for display */
-		get extensions() {
-			return this.allowedExtensions.join(', ').toUpperCase();
-		}
-
-		emitChangeEvent() {
-			if (!this.error) {
-				// Take the first file in single mode
-				const eventValue = this.multiple ? this.files : this.files[0];
-
-				this.$emit('change', eventValue);
-			}
-
-			// Reset file input
-			// Do after everything for IE
-			this.$refs.vdInputEl.value = '';
-		}
-
-		/** This function is executed when content is dropped on the component */
-		dropHandler(e: DragEvent) {
-			this.selfReset();
-
-			const data = e.dataTransfer;
-
-			if (!data) {
-				return;
-			}
-
-			const files = data.items || data.files;
-
-			// Stop if multiple files selected in single mode
-			if (this.ifTooManyFiles(files)) {
-				return;
-			}
-
-			if (data.items) {
-				// Use DataTransferItemList interface to access the file(s)
-				for (let i = 0; i < data.items.length; i++) {
-					// If dropped items aren't files, reject them
-					if (data.items[i].kind === 'file') {
-						const file = data.items[i].getAsFile();
-
-						if (!file) {
-							return;
-						}
-
-						this.validateFile(file);
-					}
-				}
-			} else {
-				// Use DataTransfer interface to access the file(s)
-				for (let i = 0; i < data.files.length; i++) {
-					this.validateFile(data.files[i]);
-				}
-			}
-
-			this.emitChangeEvent();
-		}
-
-		/** This function is executed when after a manual file selection */
-		inputValueChanged(event: HTMLInputEvent) {
-			if (!event.target) {
-				return;
-			}
-
-			const files = event.target.files;
-			this.selfReset();
-
-			// Don't do anything if no file selected
-			if (!files || !files.length) {
-				return;
-			}
-
-			// Stop if multiple files selected in single mode
-			if (this.ifTooManyFiles(files)) {
-				return;
-			}
-
-			for (let i = 0; i < files.length; i++) {
-				this.validateFile(files[i]);
-			}
-
-			this.emitChangeEvent();
-		}
-
-		ifTooManyFiles(files: FileList | DataTransferItemList) {
-			// If not in multiple mode and more than one file,
-			// return error
-			if (!this.multiple && files.length > 1) {
-				this.$emit('error', {
-					file: files,
-					code: ErrorCodes.MULTIPLE_FILES_SELECTED
-				} as ErrorEvent);
-
-				return true;
-			}
-
-			return false;
-		}
-
-		/** Compute maximum size to human readable */
-		get maxSizeReadable() {
-			return calcHumanFileSize(this.fileSizeMax, this.fileSizeUnits);
-		}
-
-		get computedAccept(): string {
-			// Property overrides the default behavior
-			if (this.accept) {
-				return this.accept;
-			}
-
-			const accept: string[] = [];
-
-			// Calc the accept="" string from the allowed extensions
-			this.allowedExtensions.forEach((type: string) => {
-				accept.push(`.${type}`);
-			});
-
-			// The result, eg. ".pdf,.jpeg,.jpg,.png"
-			return accept.join(',');
-		}
-
-		/** Expose retry function which clicks on the input */
-		public retry() {
-			this.$refs.vdInputEl.click();
-		}
 	}
 </script>
 
 <style lang="scss" scoped>
-	.file-upload {
+	.vd-file-upload {
 		width: 100%;
 		max-width: 550px;
 		cursor: pointer;
@@ -417,13 +190,13 @@
 		}
 	}
 
-	.file-upload-placeholder {
+	.vd-file-upload-placeholder {
 		display: flex;
 		align-items: center;
 		flex-direction: column;
 	}
 
-	.file-upload-input {
+	.vd-file-upload-input {
 		position: absolute;
 		width: 1px;
 		height: 1px;
@@ -432,7 +205,7 @@
 		clip: rect(1px, 1px, 1px, 1px);
 	}
 
-	.file-upload-btn {
+	.vd-file-upload-btn {
 		border-radius: 4px;
 		transition: background .25s;
 	}
