+++ packages/vue-dot/src/patterns/FileUpload/config.ts
@@ -1,0 +1,6 @@
+export const config = {
+	uploadIcon: {
+		size: 40,
+		color: 'primary'
+	}
+};
