+++ packages/vue-dot/src/patterns/FileUpload/types.d.ts
@@ -2,3 +2,7 @@ export interface ErrorEvent {
 	file: File | DataTransferItemList;
 	code: string;
 }
+
+export interface HTMLInputEvent extends Event {
+	target: HTMLInputElement & EventTarget;
+}
